[New] `AsyncMultiTreeModel` untested code
