+++ src/base/browser/basic/tree/asyncMultiTree.ts
@@ -10,8 +10,7 @@ import { Register } from "src/base/common/event";
 import { Weakmap } from "src/base/common/map";
 import { IScrollEvent } from "src/base/common/scrollable";
 import { ITreeNode } from "src/base/browser/basic/tree/tree";
-import { Iterable } from "src/base/common/iterable";
-import { isPromise } from "util/types";
+import { AsyncMultiTreeModel, IAsyncMultiTreeModel } from "src/base/browser/basic/tree/asyncMultiTreeModel";
 
 /**
  * Provides functionality to determine the children stat of the given data.
@@ -49,9 +48,9 @@ export interface IAsyncTreeNode<T> {
  */
 export class AsyncNodeConverter<T, TFilter> implements ITreeNode<T, TFilter> {
 
-    constructor(private readonly _node: ITreeNode<IAsyncTreeNode<T>, TFilter>) {}
+    constructor(private readonly _node: ITreeNode<IAsyncTreeNode<T> | null, TFilter>) {}
 
-    get data(): T { return this._node.data.data; }
+    get data(): T { return this._node.data!.data; }
     get parent(): ITreeNode<T, TFilter> | null { return this._node.parent ? new AsyncNodeConverter(this._node.parent) : null; }
     get children(): ITreeNode<T, TFilter>[] { return this._node.children.map(child => new AsyncNodeConverter(child)); }
     get depth(): number { return this._node.depth; }
@@ -101,7 +100,7 @@ export interface IAsyncMultiTree<T, TFilter> {
 }
 
 /** EXPORT FOR MODULE USAGE, DO NOT USE DIRECTLY. */
-export type AsyncWeakMap<T, TFilter> = Weakmap<ITreeNode<IAsyncTreeNode<T>, TFilter>, ITreeNode<T, TFilter>>;
+export type AsyncWeakMap<T, TFilter> = Weakmap<ITreeNode<IAsyncTreeNode<T> | null, TFilter>, ITreeNode<T, TFilter>>;
 
 /**
  * @class Wraps a {@link IMultiTree}, // TODO
@@ -118,25 +117,8 @@ export class AsyncMultiTree<T, TFilter> implements IAsyncMultiTree<T, TFilter>, 
     private readonly _disposables: DisposableManager;
 
     protected readonly _tree: IMultiTree<IAsyncTreeNode<T>, TFilter>;
-    protected readonly _root: IAsyncTreeNode<T>;
+    protected readonly _model: IAsyncMultiTreeModel<T, TFilter>;
     
-    private readonly _nodes: Map<T | null, IAsyncTreeNode<T>>;
-    private readonly _nodemap: AsyncWeakMap<T, TFilter> = new Weakmap(node => new AsyncNodeConverter(node));
-
-    private readonly _childrenProvider: IAsyncChildrenProvider<T>;
-
-    /**
-     * Storing the ongoing {@link Promise} when fetching the children stat of 
-     * the corresponding async tree node.
-     */
-    private readonly _statFetching: Map<IAsyncTreeNode<T>, Promise<Iterable<T>>>;
-    
-    /**
-     * Storing the ongoing {@link Promise} when refreshing the async tree node and
-     * all its descendants.
-     */
-    private readonly _nodeRefreshing: Map<IAsyncTreeNode<T>, Promise<void>>;
-
     // [constructor]
 
     constructor(
@@ -147,16 +129,11 @@ export class AsyncMultiTree<T, TFilter> implements IAsyncMultiTree<T, TFilter>, 
     ) {
         this._disposables = new DisposableManager();
 
-        this._tree = this.__createTree(container, renderers, itemProvider);
-        this._root = this.__createAsyncTreeNode(undefined!, null);
-        this._nodes = new Map();
-        this._nodes.set(null, this._root);
+        const unwrapper: AsyncWeakMap<T, TFilter> = new Weakmap(node => new AsyncNodeConverter(node));
 
-        this._childrenProvider = childrenProvider;
-
-        this._statFetching = new Map();
-        this._nodeRefreshing = new Map();
-
+        this._tree = this.__createTree(container, renderers, itemProvider, unwrapper);
+        this._model = this.__createModel(this._tree, childrenProvider, unwrapper);
+        
     }
 
     // [event]
@@ -176,7 +153,7 @@ export class AsyncMultiTree<T, TFilter> implements IAsyncMultiTree<T, TFilter>, 
 
     // [method]
 
-    public async refresh(data: T = this._root.data): Promise<void> {
+    public async refresh(data: T = this._model.root): Promise<void> {
         this.__refresh(data);
     }
 
@@ -192,62 +169,24 @@ export class AsyncMultiTree<T, TFilter> implements IAsyncMultiTree<T, TFilter>, 
      private __createTree(
         container: HTMLElement,
         renderers: ITreeListViewRenderer<T, TFilter, any>[],
-        itemProvider: IListItemProvider<T>
+        itemProvider: IListItemProvider<T>,
+        unwrapper: AsyncWeakMap<T, TFilter>
     ): MultiTree<IAsyncTreeNode<T>, TFilter> 
     {
         // convert the arguments into correct type (wrappers kind of stuff)
-        const asyncRenderers = renderers.map(r => new AsyncTreeRenderer(r, this._nodemap));
+        const asyncRenderers = renderers.map(r => new AsyncTreeRenderer(r, unwrapper));
         const asyncProvider = new composedItemProvider<T, IAsyncTreeNode<T>>(itemProvider);
 
         return new MultiTree<IAsyncTreeNode<T>, TFilter>(container, asyncRenderers, asyncProvider, {});
     }
 
-    /**
-     * @description Given the data, returns the corresponding async node.
-     * @param data The given data.
-     */
-    private __getAsyncNode(data: T): IAsyncTreeNode<T> {
-        
-        // given data, gets corresponding async node.
-        const node = this._nodes.get(data === this._root.data ? null : data);
-
-        if (node === undefined) {
-            throw new Error('async node is not founded');
-        }
-
-        return node;
+    private __createModel(
+        tree: IMultiTree<IAsyncTreeNode<T>, TFilter>,
+        childrenProvider: IAsyncChildrenProvider<T>,
+        unwrapper: AsyncWeakMap<T, TFilter>
+    ): IAsyncMultiTreeModel<T, TFilter> {
+        return new AsyncMultiTreeModel<T, TFilter>(tree, childrenProvider, unwrapper);
     }
-
-    /**
-     * @description Determines if the given node is an ancestor of the other.
-     * @param node The given node.
-     * @param other The other node.
-     */
-    private __isAncestor(node: IAsyncTreeNode<T>, other: IAsyncTreeNode<T>): boolean {
-        if (other.parent === null) {
-            return false;
-        } 
-        
-        else if (other.parent === node) {
-            return true;
-        }
-
-        return this.__isAncestor(node, other.parent);
-    }
-
-    /**
-     * @description Helper function for fast creating a {@link IAsyncTreeNode}.
-     */
-    private __createAsyncTreeNode(data: T, parent: IAsyncTreeNode<T> | null): IAsyncTreeNode<T> {
-        return {
-            data: data,
-            parent: parent,
-            children: [],
-            refreshing: null
-        };
-    }
-    
-    // [`refresh()` implementation]
 
     /**
      * @description Auxiliary method for `refresh()`.
@@ -256,187 +195,20 @@ export class AsyncMultiTree<T, TFilter> implements IAsyncMultiTree<T, TFilter>, 
     private async __refresh(data: T): Promise<void> {
         
         // wait until nothing is refreshing
-        if (this._root.refreshing) {
-            await this._root.refreshing;
+        const root = this._model.getAsyncNode(data);
+        if (root.refreshing) {
+            await root.refreshing;
         }
 
         // get corresponding async node
-        const node = this.__getAsyncNode(data);
+        const node = this._model.getAsyncNode(data);
 
         // wait until refresh the node and its descendants
-        await this.__refreshNode(node);
+        await this._model.refreshNode(node);
         
         // renders the whole view
         this.__render(node);
 
-    }
-
-    /**
-     * @description Refreshing the tree structure of the given node and all its 
-     * descendants.
-     * @param node The provided async tree node.
-     * @returns The {@link Promise} of the work.
-     */
-    private async __refreshNode(node: IAsyncTreeNode<T>): Promise<void> {
-
-        /**
-         * If any ongoing refreshing node has a connection with the current node 
-         * (equal, or either one is the ancestor of the other), we have to wait 
-         * until it has been done, then we retry the method.
-         */
-        for (const [other, refreshing] of this._nodeRefreshing) {
-            if (node === other || 
-                this.__isAncestor(node, other) || 
-                this.__isAncestor(other, node)) 
-            {
-                return refreshing.then(() => this.__refreshNode(node));
-            }
-        }
-
-        /**
-         * Reaching here meaning the given node and all its descendants are all 
-         * settled. We may be ready for refreshing.
-         */
-        return this.__refreshNodeAndChildren(node);
-    }
-
-    /**
-     * @description Labeling the given node is refreshing, then refreshing the 
-     * tree structure of the given node and all its descendants.
-     * @param node The given node.
-     * @returns The {@link Promise} of the work.
-     */
-    private async __refreshNodeAndChildren(node: IAsyncTreeNode<T>): Promise<void> {
-
-        // mark the current node is refreshing
-        let finishRefresh: () => void;
-        node.refreshing = new Promise((resolve, reject) => finishRefresh = resolve );
-        this._nodeRefreshing.set(node, node.refreshing);
-
-        // remove the mark once it finished
-        node.refreshing.finally(() => {
-            node.refreshing = null;
-            this._nodeRefreshing.delete(node);
-        });
-
-        try {
-
-            // wait until finish refreshing all its direct children
-            const children = await this.__refreshChildren(node);
-
-            /**
-             * we continue the same work to the direct children, refreshing the 
-             * other descendants.
-             */
-            await Promise.all(children.map(child => this.__refreshNodeAndChildren(child)));
-
-        } finally {
-            /**
-             * function will do nothing except marking the current node 
-             * refreshing state is finished.
-             */
-            finishRefresh!();
-        }
-    }
-
-    /**
-     * @description Will fetching the updated children of the given node and 
-     * refresh the tree structure by the children.
-     * @param node The given async tree node.
-     */
-    private async __refreshChildren(node: IAsyncTreeNode<T>): Promise<IAsyncTreeNode<T>[]> {
-
-        let childrenPromise: Promise<Iterable<T>>;
-        const hasChildren = this._childrenProvider.hasChildren(node.data);
-
-        if (hasChildren === false) {
-            // since the current node is a leaf, we return nothing.
-            childrenPromise = Promise.resolve(Iterable.empty());
-        }
-
-        else {
-            /**
-             * @note We may set a timer instead of choose `await`. Once timed out,
-             * we mark this node as `slow` state, and fires the event to tell 
-             * everybody. One of the usage of this could be rendering `loading` 
-             * animation. For now, I choose `await` for simplicity.
-             */
-            const children = this.__getChildren(node);
-            childrenPromise = Promise.resolve(children);
-        }
-
-        try {
-            const children = await childrenPromise;
-            return this.__setChildren(node, children);
-        } 
-        
-        catch (err) {
-            throw err;
-        } 
-        
-        finally {
-            // do nothing for now
-        }
-    }
-
-    /**
-     * @description Given the tree node, returns the newest children of the node.
-     * @param node The provided async tree node.
-     */
-    private __getChildren(node: IAsyncTreeNode<T>): Promise<Iterable<T>> | Iterable<T> {
-        let refreshing = this._statFetching.get(node);
-
-        // since the node is already fetching, we do nothing and return the same promise.
-        if (refreshing) {
-            return refreshing;
-        }
-
-        const children = this._childrenProvider.getChildren(node.data);
-        
-        if (isPromise(children)) {
-            this._statFetching.set(node, children);
-            return children.finally(() => this._statFetching.delete(node));
-        }
-
-        else {
-            return children;
-        }
-    }
-
-    /**
-     * @description Updates the given children to the provided async tree node.
-     * @param node The provided tree node.
-     * @param childrenIterable The direct children of the given node.
-     * @returns After the children were inserted, returns a iterable of the new 
-     * children tree nodes.
-     */
-    private __setChildren(node: IAsyncTreeNode<T>, childrenIterable: Iterable<T>): IAsyncTreeNode<T>[] {
-        
-        const children = [...childrenIterable];
-
-        // corner case check, performance improvement
-        if (children.length === 0 && node.children.length === 0) {
-            return [];
-        }
-
-        // create async tree node for each child
-        const childrenNodes = children.map<IAsyncTreeNode<T>>(child => {
-            return this.__createAsyncTreeNode(child, node);
-        });
-
-        // update new children mapping
-        for (const newChild of childrenNodes) {
-            this._nodes.set(newChild.data, newChild);
-        }
-
-        // delete the old children mapping
-        for (const oldChild of node.children) {
-            this._nodes.delete(oldChild.data);
-        }
-        
-        node.children.splice(0, node.children.length, ...childrenNodes);
-
-        return childrenNodes;
     }
 
     /**
