+++ src/base/browser/basic/tree/asyncMultiTreeModel.ts
@@ -1,0 +1,318 @@
+import { AsyncWeakMap, IAsyncChildrenProvider, IAsyncTreeNode } from "src/base/browser/basic/tree/asyncMultiTree";
+import { IMultiTree } from "src/base/browser/basic/tree/multiTree";
+import { ITreeModel, ITreeNode } from "src/base/browser/basic/tree/tree";
+import { Iterable } from "src/base/common/iterable";
+import { isPromise } from "util/types";
+
+
+export interface IAsyncMultiTreeModel<T, TFilter> extends ITreeModel<T, TFilter, T> {
+    
+    /**
+     * @description Refreshing the tree structure of the given node and all its 
+     * descendants.
+     * @param node The provided async tree node.
+     * @returns The {@link Promise} of the work.
+     */
+    refreshNode(node: IAsyncTreeNode<T>): Promise<void>;
+
+    /**
+     * @description Given the data, returns the corresponding {@link IAsyncTreeNode}.
+     * @param data The provided data.
+     */
+    getAsyncNode(data: T): IAsyncTreeNode<T>;
+
+}
+
+export class AsyncMultiTreeModel<T, TFilter> implements IAsyncMultiTreeModel<T, TFilter> {
+
+    // [field]
+
+    protected readonly _root: IAsyncTreeNode<T>;
+
+    private readonly _tree: IMultiTree<IAsyncTreeNode<T>, TFilter>;
+    private readonly _nodes: Map<T | null, IAsyncTreeNode<T>>;
+    private readonly _unwrapper: AsyncWeakMap<T, TFilter>;
+    
+    private readonly _childrenProvider: IAsyncChildrenProvider<T>;
+
+    /**
+     * Storing the ongoing {@link Promise} when fetching the children stat of 
+     * the corresponding async tree node.
+     */
+    private readonly _statFetching: Map<IAsyncTreeNode<T>, Promise<Iterable<T>>>;
+    
+    /**
+     * Storing the ongoing {@link Promise} when refreshing the async tree node and
+     * all its descendants.
+     */
+    private readonly _nodeRefreshing: Map<IAsyncTreeNode<T>, Promise<void>>;
+
+    // [constructor]
+
+    constructor(
+        tree: IMultiTree<IAsyncTreeNode<T>, TFilter>,
+        childrenProvider: IAsyncChildrenProvider<T>,
+        nodemap: AsyncWeakMap<T, TFilter>
+    ) {
+        
+        this._root = this.__createAsyncTreeNode(undefined!, null);
+        this._tree = tree;
+        this._nodes = new Map();
+        this._nodes.set(null, this._root);
+        this._unwrapper = nodemap;
+
+        this._childrenProvider = childrenProvider;
+
+        this._statFetching = new Map();
+        this._nodeRefreshing = new Map();
+
+    }
+
+    // [method]
+
+    get root() { return this._root.data; }
+
+    public async refreshNode(node: IAsyncTreeNode<T>): Promise<void> {
+        /**
+         * If any ongoing refreshing node has a connection with the current node 
+         * (equal, or either one is the ancestor of the other), we have to wait 
+         * until it has been done, then we retry the method.
+         */
+         for (const [other, refreshing] of this._nodeRefreshing) {
+            if (node === other || 
+                this.__isAncestor(node, other) || 
+                this.__isAncestor(other, node)) 
+            {
+                return refreshing.then(() => this.refreshNode(node));
+            }
+        }
+
+        /**
+         * Reaching here meaning the given node and all its descendants are all 
+         * settled. We may be ready for refreshing.
+         */
+        return this.__refreshNodeAndChildren(node);
+    }
+
+    public getAsyncNode(data: T): IAsyncTreeNode<T> {
+        return this.__getAsyncNode(data);        
+    }
+
+    public getNode(data: T): ITreeNode<T, TFilter> {
+        const asyncNode = this.__getAsyncNode(data);
+        const node = this._tree.getNode(asyncNode === this._root ? null : asyncNode);
+        return this._unwrapper.map(node);
+    }
+
+    public hasNode(data: T): boolean {
+        const asyncNode = this.__getAsyncNode(data);
+        return this._tree.hasNode(asyncNode === this._root ? null : asyncNode);
+    }
+
+    public getRoot(): ITreeNode<T, TFilter> {
+        const node = this._tree.getNode(this._root);
+        return this._unwrapper.map(node);
+    }
+
+    public isCollapsible(data: T): boolean {
+        const asyncNode = this.__getAsyncNode(data);
+        return this._tree.isCollapsible(asyncNode);
+    }
+
+    public isCollapsed(data: T): boolean {
+        const asyncNode = this.__getAsyncNode(data);
+        return this._tree.isCollapsed(asyncNode);
+    }
+
+    public rerender(data: T): void {
+        const asyncNode = this.__getAsyncNode(data);
+        this._tree.rerender(asyncNode);
+    }
+
+    // [private helper method]
+
+    /**
+     * @description Given the data, returns the corresponding async node.
+     * @param data The given data.
+     */
+     private __getAsyncNode(data: T): IAsyncTreeNode<T> {
+        
+        // given data, gets corresponding async node.
+        const node = this._nodes.get(data === this._root.data ? null : data);
+
+        if (node === undefined) {
+            throw new Error('async node is not founded');
+        }
+
+        return node;
+    }
+
+    /**
+     * @description Determines if the given node is an ancestor of the other.
+     * @param node The given node.
+     * @param other The other node.
+     */
+    private __isAncestor(node: IAsyncTreeNode<T>, other: IAsyncTreeNode<T>): boolean {
+        if (other.parent === null) {
+            return false;
+        } 
+        
+        else if (other.parent === node) {
+            return true;
+        }
+
+        return this.__isAncestor(node, other.parent);
+    }
+
+    /**
+     * @description Helper function for fast creating a {@link IAsyncTreeNode}.
+     */
+    private __createAsyncTreeNode(data: T, parent: IAsyncTreeNode<T> | null): IAsyncTreeNode<T> {
+        return {
+            data: data,
+            parent: parent,
+            children: [],
+            refreshing: null
+        };
+    }
+    
+    /**
+     * @description Labeling the given node is refreshing, then refreshing the 
+     * tree structure of the given node and all its descendants.
+     * @param node The given node.
+     * @returns The {@link Promise} of the work.
+     */
+    private async __refreshNodeAndChildren(node: IAsyncTreeNode<T>): Promise<void> {
+
+        // mark the current node is refreshing
+        let finishRefresh: () => void;
+        node.refreshing = new Promise((resolve, reject) => finishRefresh = resolve );
+        this._nodeRefreshing.set(node, node.refreshing);
+
+        // remove the mark once it finished
+        node.refreshing.finally(() => {
+            node.refreshing = null;
+            this._nodeRefreshing.delete(node);
+        });
+
+        try {
+
+            // wait until finish refreshing all its direct children
+            const children = await this.__refreshChildren(node);
+
+            /**
+             * we continue the same work to the direct children, refreshing the 
+             * other descendants.
+             */
+            await Promise.all(children.map(child => this.__refreshNodeAndChildren(child)));
+
+        } finally {
+            /**
+             * function will do nothing except marking the current node 
+             * refreshing state is finished.
+             */
+            finishRefresh!();
+        }
+    }
+
+    /**
+     * @description Will fetching the updated children of the given node and 
+     * refresh the tree structure by the children.
+     * @param node The given async tree node.
+     */
+    private async __refreshChildren(node: IAsyncTreeNode<T>): Promise<IAsyncTreeNode<T>[]> {
+
+        let childrenPromise: Promise<Iterable<T>>;
+        const hasChildren = this._childrenProvider.hasChildren(node.data);
+
+        if (hasChildren === false) {
+            // since the current node is a leaf, we return nothing.
+            childrenPromise = Promise.resolve(Iterable.empty());
+        }
+
+        else {
+            /**
+             * @note We may set a timer instead of choose `await`. Once timed out,
+             * we mark this node as `slow` state, and fires the event to tell 
+             * everybody. One of the usage of this could be rendering `loading` 
+             * animation. For now, I choose `await` for simplicity.
+             */
+            const children = this.__getChildren(node);
+            childrenPromise = Promise.resolve(children);
+        }
+
+        try {
+            const children = await childrenPromise;
+            return this.__setChildren(node, children);
+        } 
+        
+        catch (err) {
+            throw err;
+        } 
+        
+        finally {
+            // do nothing for now
+        }
+    }
+
+    /**
+     * @description Given the tree node, returns the newest children of the node.
+     * @param node The provided async tree node.
+     */
+    private __getChildren(node: IAsyncTreeNode<T>): Promise<Iterable<T>> | Iterable<T> {
+        let refreshing = this._statFetching.get(node);
+
+        // since the node is already fetching, we do nothing and return the same promise.
+        if (refreshing) {
+            return refreshing;
+        }
+
+        const children = this._childrenProvider.getChildren(node.data);
+        
+        if (isPromise(children)) {
+            this._statFetching.set(node, children);
+            return children.finally(() => this._statFetching.delete(node));
+        }
+
+        else {
+            return children;
+        }
+    }
+
+    /**
+     * @description Updates the given children to the provided async tree node.
+     * @param node The provided tree node.
+     * @param childrenIterable The direct children of the given node.
+     * @returns After the children were inserted, returns a iterable of the new 
+     * children tree nodes.
+     */
+    private __setChildren(node: IAsyncTreeNode<T>, childrenIterable: Iterable<T>): IAsyncTreeNode<T>[] {
+        
+        const children = [...childrenIterable];
+
+        // corner case check, performance improvement
+        if (children.length === 0 && node.children.length === 0) {
+            return [];
+        }
+
+        // create async tree node for each child
+        const childrenNodes = children.map<IAsyncTreeNode<T>>(child => {
+            return this.__createAsyncTreeNode(child, node);
+        });
+
+        // update new children mapping
+        for (const newChild of childrenNodes) {
+            this._nodes.set(newChild.data, newChild);
+        }
+
+        // delete the old children mapping
+        for (const oldChild of node.children) {
+            this._nodes.delete(oldChild.data);
+        }
+        
+        node.children.splice(0, node.children.length, ...childrenNodes);
+
+        return childrenNodes;
+    }
+
+}
